Danny/kernel 565 docs add retry logic callout to file io code snippet in docs (#123)
* Add logic to wait for file's existence

Added logic to the file download code snippets to ensure that after the download completes and we receive that CDP event, we still are then doing a while loop to check for the actual existence of the file in Kernel's file system before proceeding to read the file.

Also updated the Python example to be a working example using Kernel's synchronous SDK calls.

* Handle Cancel State in Snippet

Added conditional logic to handle download cancel state before waiting for the file to exist. Also added print statement to let the user know that the Kernel browser deleted successfully.